==> 8_Model_LightGBM.ipynb <==
import lightgbm as lgb

import pandas as pd
from sklearn.model_selection import train_test_split
import numpy as np

from sklearn.metrics import mean_squared_error
from sklearn.metrics import r2_score
from sklearn.metrics import explained_variance_score

# Transaction Level

#Get data
raw_trans_csv = pd.read_csv('../Data/train_trans.csv', index_col=0)

classification_labels = raw_trans_csv.iloc[:,1] #Unused here - regression only
regression_labels = raw_trans_csv.iloc[:,2]
transaction_data = raw_trans_csv.iloc[:,3:]

#Split test and train
X_train_1, X_test, y_train_1, y_test = train_test_split(transaction_data, regression_labels, test_size=0.33)

X_train, X_valid, y_train, y_valid = train_test_split(X_train_1, y_train_1, test_size=0.33)

#Create lgb datasets
train_data = lgb.Dataset(X_train, label=y_train)
test_data = lgb.Dataset(X_test, label=y_test)

## Hyperparameter selection

Search Space:

num_leaves : 31 (default), 50, 100

learning_rate: 0.1 (default), 0.05

max_bin: 255 (default), 300

Total iterations required:  12

### Manual Grid Search

#All params
param1 = {'num_leaves':31, 'learning_rate':0.1, 'max_bin':255,'objective':'rmse'}
param2 = {'num_leaves':50, 'learning_rate':0.1, 'max_bin':255,'objective':'rmse'}
param3 = {'num_leaves':100, 'learning_rate':0.1, 'max_bin':255,'objective':'rmse'}
param4 = {'num_leaves':31, 'learning_rate':0.05, 'max_bin':255,'objective':'rmse'}
param5 = {'num_leaves':50, 'learning_rate':0.05, 'max_bin':255,'objective':'rmse'}
param6 = {'num_leaves':100, 'learning_rate':0.05, 'max_bin':255,'objective':'rmse'}
param7 = {'num_leaves':31, 'learning_rate':0.1, 'max_bin':300,'objective':'rmse'}
param8 = {'num_leaves':50, 'learning_rate':0.1, 'max_bin':300,'objective':'rmse'}
param9 = {'num_leaves':100, 'learning_rate':0.1, 'max_bin':300,'objective':'rmse'}
param10 = {'num_leaves':31, 'learning_rate':0.05, 'max_bin':300,'objective':'rmse'}
param11 = {'num_leaves':50, 'learning_rate':0.05, 'max_bin':300,'objective':'rmse'}
param12 = {'num_leaves':100, 'learning_rate':0.05, 'max_bin':300,'objective':'rmse'}

#Custom code for grid search
def train_model_print_metrics(param_dict, train_data, X_test, y_test):
    model = lgb.train(param_dict, train_data)
    ypred = model.predict(X_test)
    
    rmse = mean_squared_error(y_test,ypred)**(1/2)
    psuedo_r2 = r2_score(y_test, ypred)
    explained_variance = explained_variance_score(y_test,ypred)
    print("RMSE: "+str(rmse)+"\n")
    print("Psuedo R2: "+str(psuedo_r2)+"\n")
    print("Explained Variance: "+str(explained_variance)+"\n")
    
    return [rmse, psuedo_r2, explained_variance]

metrics_grid = []

metrics_grid.append(train_model_print_metrics(param1, train_data, X_valid, y_valid))

metrics_grid.append(train_model_print_metrics(param2, train_data, X_valid, y_valid))

metrics_grid.append(train_model_print_metrics(param3, train_data, X_valid, y_valid))

metrics_grid.append(train_model_print_metrics(param4, train_data, X_valid, y_valid))

metrics_grid.append(train_model_print_metrics(param5, train_data, X_valid, y_valid))

metrics_grid.append(train_model_print_metrics(param6, train_data, X_valid, y_valid))

metrics_grid.append(train_model_print_metrics(param7, train_data, X_valid, y_valid))

metrics_grid.append(train_model_print_metrics(param8, train_data, X_valid, y_valid))

metrics_grid.append(train_model_print_metrics(param9, train_data, X_valid, y_valid))

metrics_grid.append(train_model_print_metrics(param10, train_data, X_valid, y_valid))

metrics_grid.append(train_model_print_metrics(param11, train_data, X_valid, y_valid))

metrics_grid.append(train_model_print_metrics(param12, train_data, X_valid, y_valid))

metrics_grid

Lowest RMSE: Parameters 5 and 11

Highest Psuedo R2: Parameters 5 and 11

Highest Explained Variance: Parameters 5 and 11

I'm choosing p10 as my final model: 

num_leaves: 31

learning_rate: 0.05

max_bin: 300



## Final model fit and metrics

#Fit final model, predict, metrics
train_model_print_metrics(param11, train_data, X_test, y_test)



Overall, this model is struggling - the negative explained variance / r2 is an indicater that this is actually doing worse than a mean-prediction

# User Level

#Get data, split test and train
raw_user_csv = pd.read_csv('../Data/train_user.csv', index_col=0)

user_classification_labels = raw_user_csv.iloc[:,1] #Unused here - regression only
user_regression_labels = raw_user_csv.iloc[:,2]
user_data = raw_user_csv.iloc[:,3:]

#Split test and train
X_train_1, X_test, y_train_1, y_test = train_test_split(user_data, user_regression_labels, test_size=0.33)

X_train, X_valid, y_train, y_valid = train_test_split(X_train_1, y_train_1, test_size=0.33)

#Create lgb datasets
train_data = lgb.Dataset(X_train, label=y_train)
test_data = lgb.Dataset(X_test, label=y_test)

## Hyperparameter Selection

Search Space:

num_leaves : 31 (default), 50, 100

learning_rate: 0.1 (default), 0.05

max_bin: 255 (default), 300

Total iterations required:  12

I will be using the same param dicts and helper function from the transactions data

metrics_grid_user = []

metrics_grid_user.append(train_model_print_metrics(param1, train_data, X_valid, y_valid))

metrics_grid_user.append(train_model_print_metrics(param2, train_data, X_valid, y_valid))

metrics_grid_user.append(train_model_print_metrics(param3, train_data, X_valid, y_valid))

metrics_grid_user.append(train_model_print_metrics(param4, train_data, X_valid, y_valid))

metrics_grid_user.append(train_model_print_metrics(param5, train_data, X_valid, y_valid))

metrics_grid_user.append(train_model_print_metrics(param6, train_data, X_valid, y_valid))

metrics_grid_user.append(train_model_print_metrics(param7, train_data, X_valid, y_valid))

metrics_grid_user.append(train_model_print_metrics(param8, train_data, X_valid, y_valid))

metrics_grid_user.append(train_model_print_metrics(param9, train_data, X_valid, y_valid))

metrics_grid_user.append(train_model_print_metrics(param10, train_data, X_valid, y_valid))

metrics_grid_user.append(train_model_print_metrics(param11, train_data, X_valid, y_valid))

metrics_grid_user.append(train_model_print_metrics(param12, train_data, X_valid, y_valid))

metrics_grid_user

Lowest RMSE: Parameters 5 and 11

Highest Psuedo R2: Parameters 5 and 11

Highest Explained Variance: Parameters 5 and 11

## Final model fit and metrics

#Fit final model, predict, metrics
train_model_print_metrics(param11, train_data, X_test, y_test)


This model is much better - very low RMSE

model = lgb.train(param11, train_data)


d = dict(zip(X_train.columns.values,model.feature_importance()))
sorted(d.items(), key=lambda kv: -kv[1])
